separate root and exclusion mask models
