+++ HTML/index.html
@@ -237,23 +237,47 @@
     </div>
     
     <div class="field">
-    <label>Model to use</label>
-    <div class="ui dropdown selection" id="settings-active-model">
-      <input type="hidden" name="active-model">
-      <i class="dropdown icon"></i>
-      <div class="default text">Select a model</div>
-      <div class="menu">
-        <div class="item" data-value="1">20200527_alnus_vessels</div>
-        <div class="item" data-value="0">20200522_conifer</div>
+      <label>Segmentation model</label>
+      <div class="ui dropdown selection" id="settings-active-model">
+        <input type="hidden" name="active-model">
+        <i class="dropdown icon"></i>
+        <div class="default text">Segmentation model</div>
+        <div class="menu">
+          <!--<div class="item" data-value="1">20200527_alnus_vessels</div>
+          <div class="item" data-value="0">20200522_conifer</div>-->
+        </div>
       </div>
     </div>
+
+
+    <div class="ui form content">
+      <div class="field">
+        <label>Exclusion mask</label>
+        <div class="ui toggle checkbox" id="settings-exclusionmask-enable">
+          <input type="checkbox">
+          <label>Enable detection of foreign objects (e.g. tape)</label>
+        </div>
+      </div>
+
+    <div class="field" id="settings-exclusionmask-model-field" style="display:none">
+      <label>Exclusion mask model</label>
+      <div class="ui dropdown selection" id="settings-exclusionmask-model">
+        <input type="hidden" name="active-model">
+        <i class="dropdown icon"></i>
+        <div class="default text">Exclusion mask model</div>
+        <div class="menu">
+          <!--<div class="item" data-value="1">20200527_alnus_vessels</div>
+          <div class="item" data-value="0">20200522_conifer</div>-->
+        </div>
+      </div>
     </div>
 
+    <div class="ui divider"></div>
 
     <div class="field">
       <label>Retraining</label>
       <div class="ui toggle checkbox" id="settings-enable-retraining">
-        <input type="checkbox">
+        <input type="checkbox" >
         <label>Enable retraining</label>
       </div>
     </div>
