+++ HTML/settings.js
@@ -1,8 +1,15 @@
 //called when user clicks on save in the settings dialog
 function save_settings(_){
     global.settings.active_model = $("#settings-active-model").dropdown('get value');
+    global.settings.exmask_model = $("#settings-exclusionmask-model").dropdown('get value');
+
     $('#settings-ok-button').addClass('loading');
-    $.post(`/settings?active_model=${global.settings.active_model}`).done((x)=>{
+    var data = JSON.stringify({
+        active_model: global.settings.active_model,
+        exmask_model: global.settings.exmask_model,
+        exmask_enabled: global.settings.exmask_enabled,
+    });
+    $.post(`/settings`, data,).done((x)=>{
         $('#settings-dialog').modal('hide');
         $('#settings-ok-button').removeClass('loading');
         console.log('Settings:',x)
@@ -24,17 +31,27 @@ function on_settings(){
 
 
 function load_settings(){
+    $('#settings-exclusionmask-enable').checkbox({onChange: on_exmask_checkbox});
+
     $.get('/settings').done(function(settings){
-        global.settings.models       = settings.models;
-        global.settings.active_model = settings.active_model;
+        console.log(settings)
+        global.settings.active_model = settings.active_model
+        global.settings.exmask_model = settings.exmask_active_model
         console.log(global.settings);
 
-        models_list = []
-        for(modelname of global.settings.models)
+        var models_list = []
+        for(var modelname of settings.models)
             models_list.push({name:modelname, value:modelname, selected:(modelname==global.settings.active_model)})
         if(settings.active_model=='')
             models_list.push({name:'[UNSAVED MODEL]', value:'', selected:true})
         $("#settings-active-model").dropdown({values: models_list, showOnFocus:false })
+
+        $('#settings-exclusionmask-enable').checkbox(settings.exmask_enabled? 'check' : 'uncheck');
+        var exmaskmodels_list = []
+        for(var name of settings.exmask_models)
+            exmaskmodels_list.push({name:name, value:name, selected:(name==global.settings.exmask_model)})
+        console.log(exmaskmodels_list)
+        $("#settings-exclusionmask-model").dropdown({values: exmaskmodels_list, showOnFocus:false })
 
         var $new_name_elements = $("#settings-new-modelname-field");
         (settings.active_model=='')? $new_name_elements.show(): $new_name_elements.hide();
@@ -64,4 +81,16 @@ function on_save_model(){
       return;
     }
     $.get('/save_model', {newname:newname}).done(load_settings);
+}
+
+
+
+function on_exmask_checkbox(){
+    var enabled = $('#settings-exclusionmask-enable').checkbox('is checked');
+    global.settings.exmask_enabled = enabled;
+    if(enabled){
+        $("#settings-exclusionmask-model-field").show()
+    } else {
+        $("#settings-exclusionmask-model-field").hide()
+    }
 }
