+++ settings.json
@@ -1,1 +1,1 @@
-{"active_model": "20210617_012c_tape"}
+{"active_model": "20200430_003x_WMS", "exmask_enabled": true, "exmask_model": "20210629_014a_tape"}
